Re-structure of the file dir
