+++ README.md
@@ -1,0 +1,23 @@
+# AD - IIT
+Currently working on the AD (jupyter notebook), that is in the:
+```bash
+AD_IIT/
+├── Alzheimer\ set
+│   ├── Approach_1.py
+│   ├── Dropper.py
+│   ├── Main.py
+│   ├── SingleFocus_1.py
+│   ├── __pycache__
+│   │   ├── Dropper.cpython-36.pyc
+│   │   └── SingleFocus_1.cpython-36.pyc
+│   ├── helper.py
+│   ├── note4codebook.txt
+│   ├── workSpace_1.ipynb
+│   ├── workSpace_2.ipynb
+│   ├── workSpace_3.ipynb
+│   ├── workSpace_4.ipynb
+│   └── workSpace_5.ipynb
+└── README.md
+
+2 directories, 14 files
+```
