--- TF-learn-RA-master/README.md
@@ -1,3 +1,0 @@
-# TF-learn-RA
-Currently working on the AD (jupyter notebook), that is in the:
-> TF-learn-RA/bio-ra/Alzheimer set/.
