--- TF-learn-RA-master/bio-ra/Alzheimer set/helper.py
@@ -1,2 +1,0 @@
-import pandas as pd
-
